# shared/specs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Edge AI Platform API
  version: 1.0.0
  description: |
    REST API for the Edge AI Platform web management interface.
    Provides device management, chat, and metrics access.
  contact:
    name: VLMChat Platform
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://localhost:3000
    description: Docker development server
security:
  - bearerAuth: []
...
paths:
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login or register user
      description: Authenticate user with email. Creates new user if doesn't exist.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                name:
                  type: string
                  example: John Doe
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/devices:
    get:
      tags: [devices]
      summary: List all devices
      description: Get all edge devices registered to the current user
      operationId: listDevices
      parameters:
        - name: status
          in: query
          description: Filter devices by status
          required: false
          schema:
            type: string
            enum: [active, connected, all]
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    
    post:
      tags: [devices]
      summary: Register new device
      description: Create a new edge device registration
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '200':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/devices/{deviceId}:
    get:
      tags: [devices]
      summary: Get device by ID
      description: Retrieve a specific device by its ID
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          $ref: '#/components/responses/InternalError'
    
    delete:
      tags: [devices]
      summary: Delete device
      description: Remove a device registration
      operationId: deleteDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalError'
  /api/devices/{deviceId}/messages:
    get:
      tags: [devices]
      summary: Get device messages
      description: Retrieve MQTT messages from a specific device
      operationId: getDeviceMessages
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceMessage'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/devices/{deviceId}/snapshot:
    post:
      tags: [devices]
      summary: Request snapshot from device
      description: Request the device to capture and send a snapshot image. The snapshot will be delivered via WebSocket when ready.
      operationId: requestDeviceSnapshot
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '202':
          description: Snapshot request accepted and sent to device. Snapshot will be delivered via WebSocket.
          content:
            application/json:
              schema:
                type: object
                required:
                  - accepted
                properties:
                  accepted:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chat/messages:
    get:
      tags: [chat]
      summary: Get chat messages
      description: Retrieve chat message history
      operationId: getChatMessages
      parameters:
        - name: deviceId
          in: query
          description: Filter by device ID
          schema:
            type: string
      responses:
        '200':
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chat/message:
    post:
      tags: [chat]
      summary: Send chat message
      description: Send a message to the AI and get a response
      operationId: sendChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: "What do you see in this image?"
                deviceId:
                  type: string
                  format: uuid
                debug:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Message sent and response received
          content:
            application/json:
              schema:
                type: object
                properties:
                  userMessage:
                    $ref: '#/components/schemas/ChatMessage'
                  aiMessage:
                    $ref: '#/components/schemas/ChatMessage'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      description: Device unique identifier
      schema:
        type: string
  schemas:
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        provider:
          oneOf:
            - type: string
              enum: [google, github, email]
            - type: 'null'
        providerId:
          oneOf:
            - type: string
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    Device:
      type: object
      description: |
        Device runtime state and identity.
        Lightweight response schema for device queries.
      required: [deviceId, name, status, createdAt]
      properties:
        deviceId:
          type: string
          description: Unique device identifier
        name:
          type: string
          description: Human-friendly device name
        status:
          type: string
          enum: [connected, disconnected, connecting, error]
          description: Current connection status
        userId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
          description: Owner user ID (null if unclaimed)
        lastSeen:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: Last activity timestamp
        createdAt:
          type: string
          format: date-time
          description: Device creation timestamp
    DeviceCreate:
      description: |
        Admin device registration payload.
        Full hardware specification provided during manufacturing/provisioning.
      allOf:
        - $ref: '../schemas/register-v1.0.0.json'
    DeviceMessage:
      type: object
      required: [id, deviceId, topic, payload, createdAt]
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        topic:
          type: string
          description: MQTT topic
        payload:
          type: object
          description: Message payload
        qos:
          type: string
          enum: ['0', '1', '2']
        retained:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ChatMessage:
      type: object
      required: [id, userId, role, content, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        deviceId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        images:
          type: array
          items:
            type: string
        debug:
          oneOf:
            - type: object
            - type: 'null'
        createdAt:
          type: string
          format: date-time
...
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Authentication required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from cookie-based authentication